pyproject.toml
[project]
name = "browsecomp-plus-retrieval-purple"
version = "0.1.0"
description = "BM25 retrieval agent for the BrowseComp-Plus benchmark"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "pyserini>=0.43.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
